chore: move playground to its own package
