+++ package.json
@@ -6,13 +6,10 @@
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
   "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
     "test:unit": "vitest -r tests",
     "test:snapshot": "node tests/test-snapshot",
     "test:integration": "node tests/test-integration",
-    "format": "prettier --write \"{src,tests}/**/*.{js,ts,svelte,md}\"",
+    "format": "prettier --write \"{src,tests,playground}/**/*.{js,ts,svelte,md}\" --ignore-path \".gitignore\"",
     "prepack": "tsc"
   },
   "dependencies": {
@@ -28,14 +25,9 @@
     "typescript": "^5.1.6"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^1.4.0",
     "@types/node": "^20.4.9",
     "@types/prettier": "^2.7.3",
-    "carbon-components-svelte": "^0.79.0",
-    "carbon-preprocess-svelte": "^0.10.0",
-    "codemirror": "^5.65.14",
     "prettier-plugin-svelte": "^2.8.0",
-    "svelte-highlight": "^7.3.0",
     "vite": "^3.2.3",
     "vitest": "^0.34.1"
   },
