+++ playground/index.html
@@ -6,7 +6,7 @@
   </head>
   <body>
     <script type="module">
-      import App from "./App.svelte";
+      import App from "./src/App.svelte";
 
       const app = new App({ target: document.body });
     </script>
