+++ playground/package.json
@@ -1,0 +1,18 @@
+{
+  "type": "module",
+  "private": true,
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "@sveltejs/vite-plugin-svelte": "^2.4.5",
+    "carbon-components-svelte": "^0.79.0",
+    "carbon-preprocess-svelte": "^0.10.0",
+    "codemirror": "^5.65.14",
+    "prettier": "^2.8.8",
+    "svelte-highlight": "^7.3.0",
+    "vite": "^4.4.9"
+  }
+}
