--- render.yaml
@@ -1,6 +1,0 @@
-services:
-  - type: web
-    name: sveld
-    env: static
-    buildCommand: yarn build
-    staticPublishPath: playground/dist
